Extract info list populator to a new class
